moved parsing out to recipeParser.js module
